backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-environ
django-filter
psycopg2-binary
whitenoise[brotli]
django-cors-headers
gunicorn